Update run.sh
Submitting the pipeline script itself, instead of running it directly on the headnode, is not necessary but also not a bad practice. I do not, however, see necessity in the 50 GB (--mem=50g) of resources for this script. This is only calculating the job graph and submitting jobs to the cluster. For the same reason, the --ntasks-per-node=10 is not needed